Rename deepcell_label to backend
