+++ backend/__init__.py
@@ -7,9 +7,9 @@ from flask_compress import Compress
 from flask_cors import CORS
 from flask_dropzone import Dropzone
 
-from deepcell_label import config
-from deepcell_label.blueprints import bp
-from deepcell_label.models import db
+from . import config
+from .blueprints import bp
+from .models import db
 
 compress = Compress()  # pylint: disable=C0103
 dropzone = Dropzone()  # pylint: disable=C0103
