+++ backend/models_test.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from deepcell_label import models
-from deepcell_label.conftest import DummyLoader
+from . import models
+from .conftest import DummyLoader
 
 
 # Automatically enable transactions for all tests, without importing any extra fixtures.
