+++ backend/pytest.ini
@@ -1,6 +1,6 @@
 # Configuration of py.test
 [pytest]
-mocked-sessions=deepcell_label.models.db.session
+mocked-sessions=models.db.session
 
 addopts=-v
         --durations=20
